Add a buffer std::vector of Modifier<> for World and Entity to be able to add Modifier during the reading of modifiers_

+++ include/utility/random.hpp
@@ -7,7 +7,7 @@
 
 int randInt(int minInclusive, int maxInclusive);
 
-int normalRandFloat(float mean, float stddev);
+float normalRandFloat(float mean, float stddev);
 
 
 #endif // UT_RANDOM_HPP
